requirements.txt
asgiref==3.9.1
Django==5.2.6
sqlparse==0.5.3
tzdata==2025.2
djangorestframework
djangorestframework-simplejwt
